feat(aliases): implement aliases for commit types
This is a decent change. Now the CommitType enum is auto-generated via a
macro. This macro also implements two helper methods, and one Trait. The
helper methods return a Vec<'static str> of all "aliases" for specified
variant if called like CommitType::Variant.aliases(). Or if called on
the enum itself there the method returns ALL aliases such as
CommitType::all_aliases(). The macro also implments the
std::str::FromStr trait so you can parse from a string to a known enum
type automatically. Say we defined a "feat" or "ft" alias of variant
"Feature", if we had a string of "ft" or "feature" we could call
some_str.parse::<CommitType>().unwrap() to get a CommitType::Feature, or if it
failed to parse we get a CommitType::Unknown

Closes #3